+++ packages/nextjs/app/components/ProductSort.tsx
@@ -1,6 +1,8 @@
+"use client";
+
 import * as React from "react";
 import { useRouterQueryParams } from "utils/useRouterQueryParams";
-import { ProductSort as BaseProductSort, ProductSortProps as BaseProps } from "shared-ui";
+import { ProductSort as BaseProductSort, ProductSortProps as BaseProps } from "../ui/shared-ui";
 
 type ProductSortProps = Pick<BaseProps, "as" | "showTitle" | "title" | "selectClassName">;
