+++ packages/nextjs/components/ProductFilters/FilterGroup.tsx
@@ -1,7 +1,9 @@
+"use client";
+
 import type { FilterGroup as FilterGroupType } from "utils/filters";
 import * as React from "react";
 import { ChangeEvent } from "react";
-import { Checkbox } from "shared-ui";
+import { Checkbox } from "../../app/ui/shared-ui";
 import { useRouterQueryParams } from "utils/useRouterQueryParams";
 
 type FilterGroupProps = {
