+++ packages/nextjs/utils/useRouterQueryParams.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 
 export const useRouterQueryParams = () => {
