+++ packages/shared-ui/components/MobileNav/MobileNavContext.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 type CartContext = {
